Move ConfigLoadContext to config package
